Prepared for separation of SystemState into ROS-specific versions.
